calculate path distances for inter edges

+++ src/simulation/state/world/graph/edge.rs
@@ -10,7 +10,7 @@ pub struct Edge {
     pub node1: Node,
     pub node2: Node,
     pub level: u32,
-    pub weight: f32,
+    pub weight: u32,
     pub clearance: u32,
     pub kind: Kind,
 }
